Incremental commit, just to get everything in.
Panels Page now installs successfully, although the schema is likely going to change. The essential infrastructure for the menu system is functioning correctly. The secondary router storage mechanism, page callbacks, temporary panel data storage mechanism (for render callbacks) and a few other things need to be designed - but we're on the home stretch.